requirements.txt
opencv-python
mediapipe
numpy
pillow
torch
torchvision
